separate out table content from component
